Add demo on GOT (and PLT) analysis
There are two source code files (`main.c` and `basket.c`), a header file
(`basket.h`) and a `Makefile`. The `bascket.c` file will be compiled
into a shared library (`libbasket.so`). The `main.c` will be compiled
and linked against the shared library, resulting in an executable
`main`. We investigate the resulting files: the `main` executable and
the `libbasket.so` library.

We used `nm`, `objdump` and `readelf` for static analysis and GDB for
dynamic analysis.